Put arbitrary behind a feature flag.
