+++ lib/vrl/parser/src/ast.rs
@@ -1,4 +1,5 @@
-use arbitrary::{Arbitrary, Unstructured};
+#[cfg(feature = "fuzz")]
+use arbitrary::Arbitrary;
 use diagnostic::Span;
 use lookup::LookupBuf;
 use ordered_float::NotNan;
@@ -11,7 +12,7 @@ use std::{
     str::FromStr,
 };
 
-use crate::{arbitrary_depth::ArbitraryDepth, lex::Error};
+use crate::Error;
 
 // -----------------------------------------------------------------------------
 // node
@@ -124,18 +125,6 @@ impl<T: Hash> Hash for Node<T> {
 #[derive(PartialEq)]
 pub struct Program(pub Vec<Node<RootExpr>>);
 
-impl<'a> Arbitrary<'a> for Program {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        // limit to 15 statements to avoid overwhelming the compiler.
-        let len = usize::arbitrary(u)? % 15;
-        let statements = (0..len)
-            .map(|_| Ok(Node::new(Span::default(), RootExpr::arbitrary(u)?)))
-            .collect::<arbitrary::Result<Vec<_>>>()?;
-
-        Ok(Program(statements))
-    }
-}
-
 impl fmt::Debug for Program {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         for expr in &self.0 {
@@ -187,17 +176,6 @@ pub enum RootExpr {
     Error(Error),
 }
 
-const DEPTH: isize = 4;
-
-impl<'a> Arbitrary<'a> for RootExpr {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        Ok(RootExpr::Expr(Node::new(
-            Span::default(),
-            Expr::arbitrary_depth(u, DEPTH)?,
-        )))
-    }
-}
-
 impl fmt::Debug for RootExpr {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         use RootExpr::*;
@@ -239,33 +217,6 @@ pub enum Expr {
     Variable(Node<Ident>),
     Unary(Node<Unary>),
     Abort(Node<Abort>),
-}
-
-impl<'a> ArbitraryDepth<'a> for Expr {
-    fn arbitrary_depth(u: &mut Unstructured<'a>, depth: isize) -> arbitrary::Result<Self> {
-        match u8::arbitrary(u)? % 5 {
-            0 if depth > 0 => Ok(Expr::Op(Node::new(
-                Span::default(),
-                Op::arbitrary_depth(u, depth - 1)?,
-            ))),
-            1 if depth == DEPTH => Ok(Expr::Assignment(Node::new(
-                Span::default(),
-                Assignment::arbitrary_depth(u, depth - 1)?,
-            ))),
-            2 => Ok(Expr::Variable(Node::new(
-                Span::default(),
-                Ident::arbitrary(u)?,
-            ))),
-            3 if depth > 0 => Ok(Expr::IfStatement(Node::new(
-                Span::default(),
-                IfStatement::arbitrary_depth(u, depth - 1)?,
-            ))),
-            _ => Ok(Expr::Literal(Node::new(
-                Span::default(),
-                Literal::arbitrary(u)?,
-            ))),
-        }
-    }
 }
 
 impl fmt::Debug for Expr {
@@ -314,18 +265,6 @@ impl fmt::Display for Expr {
 
 #[derive(Clone, PartialEq, Eq, Hash)]
 pub struct Ident(pub(crate) String);
-
-impl<'a> Arbitrary<'a> for Ident {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let path = match u8::arbitrary(u)? % 3 {
-            0 => "noog",
-            1 => "flork",
-            _ => "shning",
-        };
-
-        Ok(Ident(path.to_string()))
-    }
-}
 
 impl Ident {
     pub fn new(ident: impl Into<String>) -> Self {
@@ -376,18 +315,6 @@ pub enum Literal {
     Regex(String),
     Timestamp(String),
     Null,
-}
-
-impl<'a> Arbitrary<'a> for Literal {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let choice = usize::arbitrary(u)? % 100;
-        Ok(match choice {
-            0..=5 => Literal::String(String::arbitrary(u)?),
-            // TODO Limit the size of integers and keep them positive so both VRLs can handle it.
-            6..=10 => Literal::Boolean(bool::arbitrary(u)?),
-            _ => Literal::Integer((i64::arbitrary(u)? % 100).abs()),
-        })
-    }
 }
 
 impl fmt::Display for Literal {
@@ -458,22 +385,6 @@ impl fmt::Debug for Container {
 
 #[derive(Clone, PartialEq)]
 pub struct Block(pub Vec<Node<Expr>>);
-
-impl<'a> ArbitraryDepth<'a> for Block {
-    fn arbitrary_depth(u: &mut Unstructured<'a>, depth: isize) -> arbitrary::Result<Self> {
-        let len = usize::arbitrary(u)? % 5;
-        let statements = (0..len)
-            .map(|_| {
-                Ok(Node::new(
-                    Span::default(),
-                    Expr::arbitrary_depth(u, depth - 1)?,
-                ))
-            })
-            .collect::<arbitrary::Result<Vec<_>>>()?;
-
-        Ok(Block(statements))
-    }
-}
 
 impl Block {
     pub fn into_inner(self) -> Vec<Node<Expr>> {
@@ -644,20 +555,6 @@ pub struct IfStatement {
     pub alternative: Option<Node<Block>>,
 }
 
-impl<'a> ArbitraryDepth<'a> for IfStatement {
-    fn arbitrary_depth(u: &mut Unstructured<'a>, depth: isize) -> arbitrary::Result<Self> {
-        let predicate = Predicate::arbitrary_depth(u, depth - 1)?;
-        let consequent = Block::arbitrary_depth(u, depth - 1)?;
-        let alternative = Block::arbitrary_depth(u, depth - 1)?;
-
-        Ok(IfStatement {
-            predicate: Node::new(Span::default(), predicate),
-            consequent: Node::new(Span::default(), consequent),
-            alternative: Some(Node::new(Span::default(), alternative)),
-        })
-    }
-}
-
 impl fmt::Debug for IfStatement {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match &self.alternative {
@@ -691,15 +588,6 @@ impl fmt::Display for IfStatement {
 pub enum Predicate {
     One(Box<Node<Expr>>),
     Many(Vec<Node<Expr>>),
-}
-
-impl<'a> ArbitraryDepth<'a> for Predicate {
-    fn arbitrary_depth(u: &mut Unstructured<'a>, depth: isize) -> arbitrary::Result<Self> {
-        Ok(Predicate::One(Box::new(Node::new(
-            Span::default(),
-            Expr::arbitrary_depth(u, depth - 1)?,
-        ))))
-    }
 }
 
 impl fmt::Display for Predicate {
@@ -753,34 +641,6 @@ impl fmt::Debug for Predicate {
 #[derive(Clone, PartialEq)]
 pub struct Op(pub Box<Node<Expr>>, pub Node<Opcode>, pub Box<Node<Expr>>);
 
-impl<'a> Arbitrary<'a> for Op {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let left = Expr::arbitrary_depth(u, 100)?;
-        let right = Expr::arbitrary_depth(u, 100)?;
-        let opcode = Opcode::arbitrary(u)?;
-
-        Ok(Op(
-            Box::new(Node::new(Span::default(), left)),
-            Node::new(Span::default(), opcode),
-            Box::new(Node::new(Span::default(), right)),
-        ))
-    }
-}
-
-impl<'a> ArbitraryDepth<'a> for Op {
-    fn arbitrary_depth(u: &mut Unstructured<'a>, depth: isize) -> arbitrary::Result<Self> {
-        let left = Expr::arbitrary_depth(u, depth - 1)?;
-        let right = Expr::arbitrary_depth(u, depth - 1)?;
-        let opcode = Opcode::arbitrary(u)?;
-
-        Ok(Op(
-            Box::new(Node::new(Span::default(), left)),
-            Node::new(Span::default(), opcode),
-            Box::new(Node::new(Span::default(), right)),
-        ))
-    }
-}
-
 impl fmt::Display for Op {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{} {} {}", self.0, self.1, self.2)
@@ -793,7 +653,8 @@ impl fmt::Debug for Op {
     }
 }
 
-#[derive(Arbitrary, Copy, Clone, Eq, PartialEq, Debug)]
+#[cfg_attr(feature = "fuzz", derive(Arbitrary))]
+#[derive(Copy, Clone, Eq, PartialEq, Debug)]
 pub enum Opcode {
     Mul,
     Div,
@@ -905,19 +766,8 @@ pub enum Assignment {
     // }
 }
 
-impl<'a> ArbitraryDepth<'a> for Assignment {
-    fn arbitrary_depth(u: &mut Unstructured<'a>, depth: isize) -> arbitrary::Result<Self> {
-        let target = Node::new(Span::new(0, 5), AssignmentTarget::arbitrary(u)?);
-        let op = AssignmentOp::arbitrary(u)?;
-        let expr = Box::new(Node::new(
-            Span::new(7, 10),
-            Expr::arbitrary_depth(u, depth - 1)?,
-        ));
-        Ok(Assignment::Single { target, op, expr })
-    }
-}
-
-#[derive(Clone, PartialEq, Arbitrary)]
+#[cfg_attr(feature = "fuzz", derive(Arbitrary))]
+#[derive(Clone, PartialEq)]
 pub enum AssignmentOp {
     Assign,
     Merge,
@@ -1030,20 +880,6 @@ impl fmt::Debug for AssignmentTarget {
             External(Some(path)) => write!(f, "External({})", path),
             External(_) => f.write_str("External(.)"),
         }
-    }
-}
-
-impl<'a> Arbitrary<'a> for AssignmentTarget {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let path = match u8::arbitrary(u)? % 3 {
-            0 => "noog",
-            1 => "flork",
-            _ => "shning",
-        };
-
-        Ok(AssignmentTarget::External(Some(
-            LookupBuf::from_str(path).unwrap(),
-        )))
     }
 }
